drivers: nvmem: nvmem-huk: allow to use a larger cell
Currently, the length of the HW unique key must be 16 bits.

Moreover the size of NVMEM cells can be larger than 16 bits, for
example on samad5 the SFC is using 32 bits cells.

These changes allow the use of a NVMEM cell larger to the HW
unique key length, using the first 16 bits of a cell.

Signed-off-by: Thomas Perrot <thomas.perrot@bootlin.com>